+++ includes/zuplus-functions.php
@@ -7,12 +7,14 @@ class ZU_PlusFunctions {
 
 	private static $_zufunc_instance;
 	
+	private $theme_version = null;
 	private $random_attachment_id = null;
 	private 	$advanced_style = [];
 	private $admin_style = [];
 	private $fonts = [];
 	private $copy_string = '';
 	private $cache_time = HOUR_IN_SECONDS * 12; 	// cache them for 12 hours (recommended)
+	private $debug_cache = true;
 	
 	function __construct() {
 		
@@ -32,6 +34,15 @@ class ZU_PlusFunctions {
 	
 	// Cache functions -----------------------------------------------------------]
 	
+	public function get_theme_version() {
+		
+		if(!empty($this->theme_version)) return $this->theme_version;
+		
+		$theme_info = wp_get_theme();
+		$this->theme_version = $theme_info->display('Version');
+		return $this->theme_version;
+	}
+
 	public function array_md5($array, $only_json = false) {
 		
 		// https://stackoverflow.com/questions/2254220/php-best-way-to-md5-multi-dimensional-array
@@ -41,30 +52,50 @@ class ZU_PlusFunctions {
 	    return $only_json ? json_encode($array) : md5(json_encode($array));
 	}
 
+	public function set_debug_cache($value) {
+
+		$this->debug_cache = $value;
+	}
+
 	public function create_cachekey($prefix, $array = [], $string = '') {
 
-		$string = explode(' ', $string);
 		$array = is_array($array) ? $array : [];
-		if(!empty($string)) $array[$prefix.'_md5_strings'] = $string;
-		return empty($array) ? $prefix : $prefix .'_'. $this->array_md5($array);
+		$array['zuplus_version'] = sprintf('%1$s_%2$s', ZUPLUS_VERSION, $this->get_theme_version());
+
+		if(!empty($string)) $array[$prefix.'_md5_strings'] = explode(' ', trim($string));
+
+		if($this->debug_cache) {
+			zu_write_log('Create Cachekey=', ['cache_id' => sprintf('zu_%1$s_%2$s', $prefix, $this->array_md5($array)), 'array' => $array]);
+		}
+
+		return sprintf('zu_%1$s_%2$s', $prefix, $this->array_md5($array));
 	}
 
 	public function get_cached($cache_id) {
 
 		$cached = get_transient($cache_id);
+		
+		if($this->debug_cache) zu_write_log('Get Cachekey=', ['cache_id' => $cache_id, 'cached' => $cached === false ? 'NOT FOUND' : 'found: ~'.$this->get_cached_size($cached)]);
+		
 		return $cached === false ? '' : $cached;
 	}
 
+	public function get_cached_size($cached, $formated = true) {
+		$size = is_string($cached) ? strlen($cached) : strlen(serialize($cached));
+		return $formated ? $this->format_bytes($size) : $size;
+	}
+	
 	public function set_cached($cache_id, $data, $format = false) {
 		
-		if($format == 'html') $data =  $this->minify_html($data);
-		if($format == 'css') $data =  $this->minify_css($data);
+		if($format == 'html') $data = $this->minify_html($data);
+		if($format == 'css') $data = $this->minify_css($data);
+		
+		if($this->debug_cache) zu_write_log('Set Cached=', ['cache_id' => $cache_id, 'size' => $this->get_cached_size($data)]);
 		
 		set_transient($cache_id, $data, $this->cache_time);
 	}
 	
-	
-	public function purge_transients($prefix) {			
+	public function purge_transients($prefix = 'zu_') {			
 		global $wpdb;
 	
 		// Purge all the transients associated with our prefix
@@ -79,7 +110,8 @@ class ZU_PlusFunctions {
 		
 		foreach($transients as $transient) {
 			$key = str_replace('_transient_timeout_', '', $transient);  		// Strip away the WordPress prefix in order to arrive at the transient key.
-			delete_transient( $key );														// Now that we have the key, use WordPress core to the delete the transient.
+			delete_transient($key);															// Now that we have the key, use WordPress core to the delete the transient.
+			if($this->debug_cache) zu_write_log('Deleted Cached=', $key);
 		}
 		
 		wp_cache_flush();																		// Sometimes transients are not in the DB, so we have to do this too